move ProcessedService to the service_generator.rs
